Handle incomplete JSON lines frames
It is possible for a frame coming from a FramedRead to end at a closing brace for a nested object in a JSON line. In this case, the handling for lack of newlines in some APIs in OS X will attempt to parse this as a full line and fail. This can result in very intermittent failures in high volume scenarios. The correct behavior here is to attempt to parse the line and return Ok(None) if the parsing fails. This will indicate to the FramedRead that more data is needed.